文生视频/scifi_animator/requirements.txt
opencv-python-headless
opencv-python
opencv-contrib-python
numpy==1.24.3
moviepy==1.0.3
python-dotenv==1.0.0
requests>=2.32.2
filelock
fsspec[http]>=2023.1.0
jinja2<4.0
markupsafe~=2.0
accelerate>=0.21.0
torch>=1.13.0